fixed: removed some warnings.
git-svn-id: https://xbmc.svn.sourceforge.net/svnroot/xbmc/branches/linuxport/XBMC@17274 568bbfeb-2a22-0410-94d2-cc84cf5bfa90

+++ xbmc/cores/paplayer/paplayer_linux.cpp
@@ -821,9 +821,9 @@ void PAPlayer::Seek(bool bPlus, bool bLargeStep)
   {
     float percent;
     if (bLargeStep)
-      percent = bPlus ? g_advancedSettings.m_musicPercentSeekForwardBig : g_advancedSettings.m_musicPercentSeekBackwardBig;
+      percent = bPlus ? (float)g_advancedSettings.m_musicPercentSeekForwardBig : (float)g_advancedSettings.m_musicPercentSeekBackwardBig;
     else
-      percent = bPlus ? g_advancedSettings.m_musicPercentSeekForward : g_advancedSettings.m_musicPercentSeekBackward;
+      percent = bPlus ? (float)g_advancedSettings.m_musicPercentSeekForward : (float)g_advancedSettings.m_musicPercentSeekBackward;
     seek = (__int64)(GetTotalTime64()*(GetPercentage()+percent)/100);
   }
 
@@ -961,7 +961,7 @@ bool PAPlayer::AddPacketsToStream(int stream, CAudioDecoder &dec)
     memcpy(m_pcmBuffer[stream]+m_bufferPos[stream], pcmPtr, len);
     m_bufferPos[stream] += len;
 
-    while (m_bufferPos[stream] >= m_pAudioDecoder[stream]->GetChunkLen())
+    while (m_bufferPos[stream] >= (int)m_pAudioDecoder[stream]->GetChunkLen())
     {
       int rtn = m_pAudioDecoder[stream]->AddPackets(m_pcmBuffer[stream], m_bufferPos[stream]);
       m_bufferPos[stream] -= rtn;
